+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Added `Jackson2JsonBuilder.with` method to run code on a different `Jackson2JsonBuilder` with different settings.
 - Added `Jackson2JsonBuilder.getObjectMapper` to configure the underlying `ObjectMapper`.
 - Added `Jackson2JsonBuilder(JsonFactory)` constructor to configure the `ObjectMapper` with a different factory since it can only be configured at construction.
+- Added `Jackson2JsonBuilder.fromJsonInputStream` to de-serialize JSON objects from input streams.
 - Extracted `ApiphanyJackson2Module` from `Jackson2JsonBuilder` for easier usage and added `instance` method.
 - Added `Iterables` utility interface for `Iterable` and `Collection`.
 - Added `Temporals.parseSimpleDuration` to parse durations like `1s`, `2m`, `3h` or `7d`.